api/requirements.txt
# ============================================================================
# CE365 Agent - API Requirements
# ============================================================================

# FastAPI & Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.13.1  # Database migrations

# Redis
redis==5.0.1

# Anthropic API
anthropic==0.18.0

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0

# CE365 Core Dependencies
psutil==5.9.7  # System info
rich==13.7.0  # Console UI (für CLI-Kompatibilität)

# Security (PII Detection)
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2

# Web Search (optional)
duckduckgo-search==4.1.0
